refactor : 콘솔 서비스와 웹 서비스 분리
