+++ src/main/java/racingcar/RacingCarConsoleApplication.java
@@ -1,12 +1,18 @@
 package racingcar;
 
 import racingcar.controller.RacingCarConsoleController;
+import racingcar.service.RacingCarService;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class RacingCarConsoleApplication {
 
     public static void main(String[] args) {
         try {
-            new RacingCarConsoleController().run();
+            RacingCarConsoleController racingCarConsoleController = new RacingCarConsoleController(
+                    new InputView(System.in), new OutputView(), new RacingCarService()
+            );
+            racingCarConsoleController.run();
         } catch (IllegalArgumentException exception) {
             System.out.println(exception.getMessage());
         }
