+++ src/main/java/racingcar/RacingCarWebApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class RacingCarWebApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(RacingCarWebApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(RacingCarWebApplication.class, args);
+    }
 
 }
