+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import racingcar.controller.dto.GameRequestDtoForPlays;
 import racingcar.controller.dto.GameResponseDto;
 import racingcar.controller.dto.RacingGameResultDto;
-import racingcar.service.RacingCarService;
+import racingcar.service.WebService;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -15,23 +15,23 @@ import java.util.List;
 @RestController
 public class RacingCarController {
 
-    private final RacingCarService racingCarService;
+    private final WebService webService;
 
     @Autowired
-    public RacingCarController(RacingCarService racingCarService) {
-        this.racingCarService = racingCarService;
+    public RacingCarController(WebService webService) {
+        this.webService = webService;
     }
 
     @GetMapping("/plays")
     public ResponseEntity<List<GameResponseDto>> getGameHistory() {
         return ResponseEntity.ok()
-                .body(racingCarService.getGameHistory());
+                .body(webService.getGameHistory());
     }
 
     @PostMapping("/plays")
     public ResponseEntity<GameResponseDto> playGame(@RequestBody GameRequestDtoForPlays gameRequestDtoForPlays, HttpServletResponse response) throws IOException {
-        RacingGameResultDto racingGameResultDto = racingCarService.plays(gameRequestDtoForPlays);
-        int savedGameId = racingCarService.saveGameResult(racingGameResultDto);
+        RacingGameResultDto racingGameResultDto = webService.plays(gameRequestDtoForPlays);
+        int savedGameId = webService.saveGameResult(racingGameResultDto);
         response.sendRedirect("/gameResult?savedGameId=" + savedGameId);
         return ResponseEntity.ok().build();
     }
@@ -39,7 +39,7 @@ public class RacingCarController {
     @GetMapping("/gameResult")
     public ResponseEntity<GameResponseDto> gameResult(@RequestParam int savedGameId) {
         return ResponseEntity.ok()
-                .body(racingCarService.getSavedGameById(savedGameId));
+                .body(webService.getSavedGameById(savedGameId));
     }
 
 }
